feat: Create functions that validate the form and handle the errors
The functions that validate the form, trying to find errors before the form is submitted to the server has been created. The function that handles the errors and correctly displays them have also been created. An error card has been hardcoded into the code and will show up when an error appears